requirements.txt
# ray
ray[data,train,tune,serve]==2.44.1
pynvml

# torch
torch
torchvision
torchmetrics

# jupyter
nbformat
ipykernel
ipywidgets

# upnp
miniupnpc

# fastapi
fastapi[standard]
pydantic_settings

# misc
tqdm
click
pandas
pyarrow
openpyxl
colorama
matplotlib
